backend/requirements.txt
fastapi
uvicorn

# LangChain + wrappers
langchain
langchain-community
langchain-openai
langchain-text-splitters

# Vector store + utils
chromadb
python-dotenv
pypdf
